+++ src/active-response/active_responses.h
@@ -108,6 +108,31 @@ char* get_extra_args_from_json(const cJSON *input);
  * */
 char* get_keys_from_json(const cJSON *input);
 
+/**
+ * @brief this function split string
+ * @param output_buf buffer output
+ * @param delimiter  delimiter with which want to split string
+ * @param strBefore  buffer to store split string
+ * @param strAfter   buffer to store split string
+ * @return pointer to delimiter or NULL if no found
+*/
+char* splitStrFromCharDelimiter(const char * output_buf, const char delimiter, char * strBefore, char *strAfter);
+
+/**
+ * @brief It looks for a string that matches pattern 1, if it finds it, it looks again for pattern 2, there should be spaces in the middle between the patterns.
+ * @param output_buf buffer where search
+ * @param str_pattern_1 pattern to match
+ * @param str_pattern_2 pattern to match
+ * @return 1 or 0
+ * @example output_buf -> "... Status:    Disabled ..."
+ *          isEnabledFromPattern(output_buf, "Status: ", "Enabled")
+ *            if it matches pattern 1 look for pattern 2 and if found, it returns 1
+ *          isEnabledFromPattern(output_buf, "Status: ", NULL)
+ *            find only by "Status"
+*/
+int isEnabledFromPattern(const char * output_buf, const char * str_pattern_1, const char * str_pattern_2);
+
+
 #ifndef WIN32
 
 /**
